fix(style) improve script and fix C/C++ compilation
